backend/requirements.txt
# Framework web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de datos
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Configuración y validación
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyotp==2.9.0
qrcode[pil]==7.4.2

# Redis para caché
redis==5.0.1

# IA y procesamiento
openai==1.3.7
langchain==0.0.340
tiktoken==0.5.2

# Utilidades
python-dateutil==2.8.2
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Desarrollo
black==23.11.0
isort==5.12.0
flake8==6.1.0